feat: 시간표가 없는 강의에 대한 자동 create 로직 훅으로 분리
